refactor(*): Create index.js
