+++ src/commons/components/index.js
@@ -1,0 +1,7 @@
+import Card from "./Card/Card";
+import CardList from "./CardList/CardList";
+import Footer from "./Footer/Footer";
+import Nav from "./Nav/Nav";
+import SearchBar from "./SearchBar/SearchBar";
+
+export { Card, CardList, Footer, Nav, SearchBar };
