+++ src/commons/styles/colors.scss
@@ -1,0 +1,17 @@
+:root {
+  --color-primary: #6d6afe;
+  --color-red: #ff5b56;
+  --color-black: #111322;
+  --color-white: #ffffff;
+  --color-gray5: #3e3e43;
+  --color-gray4: #9fa6b2;
+  --color-gray3: #ccd5e3;
+  --color-gray2: #e7effb;
+  --color-gray1: #f0f6ff;
+  --grey-light: #f5f5f5;
+  --gra-purpleblue-to-skyblue: linear-gradient(
+    91deg,
+    var(--color-primary) 0.12%,
+    #6ae3fe 101.84%
+  );
+}
