+++ src/commons/styles/global.scss
@@ -1,0 +1,1 @@
+@import "./colors.scss";
