+++ src/commons/styles/reset.css
@@ -1,40 +1,34 @@
 * {
   box-sizing: border-box;
   font-family: "Pretendard";
+  word-break: keep-all;
   margin: 0;
   padding: 0;
 }
 
 html,
 body {
-  word-break: keep-all;
   font-size: 62.5%;
-  font-family: "Noto Sans Korean", "Poppins", "sans-serif";
-  margin: 0;
 }
 
 a {
+  color: inherit;
+  cursor: pointer;
   text-decoration: none;
 }
 
-button {
-  cursor: pointer;
+input:focus {
+  outline: none;
 }
 
-:root {
-  --color-primary: #6d6afe;
-  --color-red: #ff5b56;
-  --color-black: #111322;
-  --color-white: #ffffff;
-  --color-gray5: #3e3e43;
-  --color-gray4: #9fa6b2;
-  --color-gray3: #ccd5e3;
-  --color-gray2: #e7effb;
-  --color-gray1: #f0f6ff;
-  --grey-light: #f5f5f5;
-  --gra-purpleblue-to-skyblue: linear-gradient(
-    91deg,
-    var(--color-primary) 0.12%,
-    #6ae3fe 101.84%
-  );
+input[type="search"]::-webkit-search-decoration,
+input[type="search"]::-webkit-search-cancel-button,
+input[type="search"]::-webkit-search-results-button,
+input[type="search"]::-webkit-search-results-decoration {
+  display: none;
+}
+
+button {
+  border: none;
+  cursor: pointer;
 }
