+++ src/pages/FolderPage/components/index.js
@@ -1,0 +1,13 @@
+import FolderHeader from "./FolderHeader/FolderHeader";
+import FolderNameButton from "./FolderNameButton/FolderNameButton";
+import Header from "./Header/Header";
+import LinkAddInput from "./LinkAddInput/LinkAddInput";
+import MobileFolderButton from "./MobileFolderButton/MobileFolderButton";
+
+export {
+  FolderHeader,
+  FolderNameButton,
+  Header,
+  LinkAddInput,
+  MobileFolderButton,
+};
