+++ src/pages/index.js
@@ -1,0 +1,5 @@
+import FolderPage from "./FolderPage/FolderPage";
+import SharedPage from "./SharedPage/SharedPage";
+import HomePage from "./HomePage/HomePage";
+
+export { FolderPage, SharedPage, HomePage };
